LayoutTests/imported/w3c/web-platform-tests/content-security-policy/generic/generic-0_2_3-expected.txt

CONSOLE MESSAGE: Refused to load http://www.localhost:8800/content-security-policy/generic/unreached.js because it does not appear in the script-src directive of the Content Security Policy.
'self' fails with a different host (including sub-host e.g. foo.com as self with content from bar.foo.com)


Harness Error (TIMEOUT), message = null

PASS Prevents access to external scripts.
NOTRUN Should fire violation events for every failed violation
